Menu i Bugfix
Dodanie opisów do badań jako pojawiające się Canvasy oraz update zmiennych w plikach
