+++ PandaVision/Assets/Scripts/Descriptions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 435fdc749a68ee542b274bebe88058a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
