+++ PandaVision/Assets/Scripts/Descriptions/Descriptions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e61808d5d6e37944fb72e65a9ec5d90e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
